refactor(sui-studio): abstract error logging and file write to reusable helpers
