+++ packages/sui-studio/bin/helpers/cli.js
@@ -1,3 +1,5 @@
+const program = require('commander')
+const colors = require('colors')
 
 /**
  * Calls sui-mono comman with sui-studio default config
@@ -23,4 +25,15 @@ function callCommand (binPath, args) {
     .on('exit', code => { code && process.exit(code) })
 }
 
-module.exports = { callCommand, callSuiMonoCommand }
+/**
+ * Shows an error in the command line and exits process
+ * @param  {String} msg
+ * @return
+ */
+const showError = (msg) => {
+  program.outputHelp(txt => colors.red(txt))
+  console.error(colors.red(msg)) // eslint-disable-line no-console
+  process.exit(1)
+}
+
+module.exports = { callCommand, callSuiMonoCommand, showError }
